Add an optional feature to enable rustls
Sometimes it's useful to use rustls instead of the default native TLS.
The reqwest crate exposes a `rustls-tls` feature that makes this easy
to enable. This patch adds a new feature to this crate that passes
`rustls-tls` through to reqwest. Like before, the default TLS stack for
a platform is used by reqwest if no override is given.